# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: EthixAI API
  description: |
    Ethical AI Governance Platform API - Real-time bias detection, SHAP explanations,
    and comprehensive monitoring for financial decision-making systems.
  version: '1.0.0'
  contact:
    name: EthixAI Support
    email: support@ethixai.com
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.ethixai.com
    description: Production
...
paths:
...
  /api/governance/approval-workflows:
    post:
      tags: [Governance]
      summary: Create approval workflow
      operationId: createWorkflow
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowInput'
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '400':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [Governance]
      summary: List approval workflows
      operationId: listWorkflows
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, completed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workflow'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /api/governance/approval-workflows/{workflowId}:
    get:
      tags: [Governance]
      summary: Get workflow details
      operationId: getWorkflow
      security:
        - bearerAuth: []
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404':
          description: Workflow not found
  /api/governance/approval-workflows/{workflowId}/approvals/{approvalRequestId}:
    post:
      tags: [Governance]
      summary: Submit approval decision
      operationId: submitApproval
      security:
        - bearerAuth: []
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
        - name: approvalRequestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approved]
              properties:
                approved:
                  type: boolean
                reason:
                  type: string
                comments:
                  type: string
      responses:
        '200':
          description: Approval recorded
        '404':
          description: Request not found
  /api/governance/compliance-signoffs:
    post:
      tags: [Governance]
      summary: Create compliance sign-off request
      operationId: createSignOff
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignOffInput'
      responses:
        '201':
          description: Sign-off request created
  /api/governance/audit-callbacks:
    post:
      tags: [Governance]
      summary: Submit audit callback
      operationId: submitAuditCallback
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                callback_type:
                  type: string
                payload:
                  type: object
      responses:
        '200':
          description: Callback processed

  # ==================== AUDIT LOGS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    WorkflowInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [model_approval, policy_change, compliance_review]
        stages:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              approvers:
                type: array
                items:
                  type: string
              required_approvals:
                type: integer
        entityType:
          type: string
        entityId:
          type: string
        entityVersion:
          type: string
    Workflow:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [pending, in_progress, approved, rejected, completed]
        stages:
          type: array
          items:
            type: object
        initiatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SignOffInput:
      type: object
      required: [type, entity_id, signee]
      properties:
        type:
          type: string
        entity_id:
          type: string
        signee:
          type: string
        comments:
          type: string

    # ==================== AUDIT SCHEMAS ====================
...
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
        total_pages:
          type: integer
        total_records:
          type: integer
        per_page:
          type: integer
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: validation_failed
            message: 'Invalid input parameters'
...